unify two kinds of dataset visualization
